fix(quality): guard primary agent Write/Execute capabilities before cross-review
CodexAgent ignores allowed_tools and runs with its configured sandbox
(e.g. danger-full-access) regardless of the Some(vec![]) passed to
run_cross_review. The existing guard only filtered the challenger;
the primary was left unchecked, allowing workspace mutation during
the background quality gate.

Fix: apply the same Write/Execute capability check to the default
primary agent. If it advertises Write or Execute, skip cross-review
with a warning (same behaviour as the existing challenger guard).

Also extract quality_trigger tests to quality_trigger_tests.rs to
keep the implementation file under the 800-line hard limit, following
the existing *_tests.rs convention in harness-server.

Bump patch version to 0.6.31.
